Feature item back

+++ backend/daoumarket/src/main/java/com/daoumarket/service/IItemService.java
@@ -15,8 +15,8 @@ public interface IItemService {
 	public ResponseEntity<BasicResponse> updateItemInfo(ItemUpdateRequest item);
 	public ResponseEntity<BasicResponse> updateItemStatus(ItemUpdateRequest item);
 	public ResponseEntity<BasicResponse> deleteItem(long itemId);
-	public ResponseEntity<BasicResponse> getAllItems(long userId);
+	public ResponseEntity<BasicResponse> getAllItems(int userId);
 	public ResponseEntity<BasicResponse> getItemsByCategory(ItemSearchRequest search);
 	public ResponseEntity<BasicResponse> getItemsByKeyword(ItemSearchRequest search);
-	public ResponseEntity<BasicResponse> getItemsByUserId(long userId);
+	public ResponseEntity<BasicResponse> getItemsByUserId(int userId);
 }
